requirements.txt
Flask~=2.0.3
PyJWT~=1.7.1
Werkzeug~=2.0.3


